+++ pkg/resource/resource.go
@@ -69,9 +69,9 @@ type TargetKind schema.GroupVersionKind
 // A CompositeKind contains the type metadata for a kind of composite resource.
 type CompositeKind schema.GroupVersionKind
 
-// A RequirementKind contains the type metadata for a kind of requirement
-// resource.
-type RequirementKind schema.GroupVersionKind
+// A CompositeClaimKind contains the type metadata for a kind of composite
+// resource claim.
+type CompositeClaimKind schema.GroupVersionKind
 
 // A LocalConnectionSecretOwner may create and manage a connection secret in its
 // own namespace.
